requirements.txt
pyyaml>=6.0
streamlit>=1.28.0
langchain-core>=0.1.0
pydantic>=2.0.0
coverage>=7.0.0
python-dotenv>=1.0.0

